Hotfix broken outlook EML attachments
When outlook attaches an EML that also contains attachments the structure gets reported properly but then later a second attachment it reported that’s within that EML as an attachment of the parent email.

This attachment is not a duplicate but some byte offset from the EML attachment (maybe? not sure yet). The encoding for this attachment is reported incorrectly causing base64 decoding to fail which causes code turning errors into exceptions to hard fail. This hotfixes the issue by attempting to save the email and then removing the stream filter if that fails.

This is NOT the right solution. What we should be doing is fixing the structure generation so the 2nd attachment is skipped (preferrable) or is reported properly.